Add Ruff, mypy and uv-based CI checks
- CI now has a quality job (Ruff lint, Ruff format check, mypy on
  Python 3.14) and a test job on Python 3.9, 3.10, 3.11 and 3.14,
  both installed with uv on current action versions
- Ruff replaces flake8; settings and pinned tool versions live in
  pyproject.toml (quality dependency group)
- Code formatted with Ruff; small lint and type-hint fixes with no
  behaviour change
- Newer type-hint syntax rules stay off while Python 3.9 is supported

+++ examples/make_sample.py
@@ -63,8 +63,8 @@ def make_rows(seed: int):
         weight = rng.uniform(*BACKGROUND)
         for _, spot_lat, spot_lon, peak, spread_km in HOT_SPOTS:
             distance_km = haversine(lat, lon, spot_lat, spot_lon) / 1000
-            weight += peak * math.exp(-(distance_km ** 2) / (2 * spread_km ** 2))
-        value = int(round(weight))
+            weight += peak * math.exp(-(distance_km**2) / (2 * spread_km**2))
+        value = round(weight)
         if value == 0 or rng.random() < DROP_RATE:
             continue  # keep some data gaps, like real data
         rows.append((cell, value))
