+++ geohash_partition/style.py
@@ -9,8 +9,19 @@ from __future__ import annotations
 from typing import Iterable, Sequence, Tuple
 
 SEQUENTIAL_BLUE: Sequence[str] = (
-    "#cde2fb", "#b7d3f6", "#9ec5f4", "#86b6ef", "#6da7ec", "#5598e7",
-    "#3987e5", "#2a78d6", "#256abf", "#1c5cab", "#184f95", "#104281", "#0d366b",
+    "#cde2fb",
+    "#b7d3f6",
+    "#9ec5f4",
+    "#86b6ef",
+    "#6da7ec",
+    "#5598e7",
+    "#3987e5",
+    "#2a78d6",
+    "#256abf",
+    "#1c5cab",
+    "#184f95",
+    "#104281",
+    "#0d366b",
 )
 AREA_STROKE = "#0d366b"
 AREA_FILL_OPACITY = 0.55
@@ -32,5 +43,5 @@ def ramp_color(value: float, low: float, high: float, ramp: Sequence[str] = SEQU
     if high <= low:
         return ramp[len(ramp) // 2]
     position = (float(value) - low) / (high - low)
-    index = min(len(ramp) - 1, max(0, int(round(position * (len(ramp) - 1)))))
+    index = min(len(ramp) - 1, max(0, round(position * (len(ramp) - 1))))
     return ramp[index]
